Move socket handling into a thunk, clean up old / bad code
