+++ web/src/api/v1/state/types.ts
@@ -2,7 +2,15 @@ import type { RipState } from "@/api/v1/rip/store";
 import type { TocState } from "@/api/v1/toc/store";
 import type { SocketProgress, SocketState } from "@/api/v1/socket/store";
 
-import type { ProgressMessageEvent } from "@/components/socket/Context";
+export type CurrentStatus = "Scanning CD-ROM devices"
+      | "Opening DVD disc"
+      | "Processing title sets"
+      | "Scanning contents"
+      | "Processing titles"
+      | "Decrypting data"
+      | "Saving all titles to MKV files"
+      | "Analyzing seamless segments"
+      | "Saving to MKV file";
 
 export type State = {
   redux: {
@@ -16,7 +24,7 @@ export type State = {
     current_progress?: SocketProgress[];
     total_progress?: SocketProgress;
 
-    current_status?: ProgressMessageEvent["name"];
+    current_status?: CurrentStatus;
     total_status?: string;
 
     rip_started?: boolean;
